+++ Vulkan/shaders/shadow.geom
@@ -19,8 +19,7 @@ void main()
 	for (int i = 0; i < gl_in.length(); i++)
 	{
 		gl_Layer = gl_InvocationID;
-		vec4 tmpPos = gl_in[i].gl_Position;
-		gl_Position = ubo.mvp[gl_InvocationID] * instancedPos * tmpPos;
+		gl_Position = ubo.mvp[gl_InvocationID] * instancedPos * gl_in[i].gl_Position;
 		EmitVertex();
 	}
 	EndPrimitive();
